backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
decouple==0.0.7
Django==4.1.7
djangorestframework==3.14.0
pycodestyle==2.10.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
